requirements.txt
PyYAML
requests==2.18.4
requests-mock==1.5.0
